requirements.txt
Pillow
tqdm
#open_clip_torch>=2.23.0
#torch==2.1.0
accelerate>=0.24.1
safetensors>=0.4.0
transformers>=4.35.0
gradio>=3.23.0
huggingface-hub
pytorch_optimizer==3.3.2
torchao
triton-windows
matplotlib
bitsandbytes
sentencepiece
timm